Selected item can now be removed from shopping cart.

+++ data/src/main/java/si/lanisnik/restaurantorder/data/repository/order/OrderCache.kt
@@ -12,7 +12,7 @@ interface OrderCache {
 
     fun clearShoppingCart(): Completable
     fun addItemToShoppingCart(item: SelectedMenuItemEntity): Completable
-    fun removeItemFromShoppingCart(id: Int): Completable
+    fun removeItemFromShoppingCart(id: Long): Completable
     fun loadShoppingCart(): Single<List<SelectedMenuItemEntity>>
 
 }
